Made pages for each catagory
